% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path_fname.R
\name{add_path_fname}
\alias{add_path_fname}
\title{Add paths and filenames to a df}
\usage{
add_path_fname(path = NULL, col_names = NULL, file_ext = NULL, ...)
}
\arguments{
\item{path}{Path to files, Default: NULL}

\item{col_names}{Names for the columns. Default: NULL}

\item{file_ext}{File extension, Default: NULL}

\item{...}{extra parameters to pass to tidyr::separate}
}
\value{
A data-frame.
}
\description{
This helper function wraps `fs` functions to add the path and
filename without extension as columns in a df.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[fs]{dir_ls}},\code{\link[fs]{path_ext_remove}},\code{\link[fs]{path_file}}
 \code{\link[dplyr]{mutate}}
}
