% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cases.R
\name{clean_cases}
\alias{clean_cases}
\title{Keep only complete cases}
\usage{
clean_cases(df, id, value)
}
\arguments{
\item{df}{A dataframe}

\item{id}{Column with the subject id key}

\item{value}{Column with the values}
}
\value{
A dataframe
}
\description{
If a participant has data for more than one session, and you want to keep only participants with complete datasets (i.e. no missing data in either condition), this function will exclude those with any missing values.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
