% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_list.R
\name{export_list}
\alias{export_list}
\title{Export a list of data frames}
\usage{
export_list(list, folder = NULL, format = "csv")
}
\arguments{
\item{list}{a list of data frames.}

\item{folder}{the path to destination.}

\item{format}{the format in which to save the data. Currently only "rds"
and "csv" (Default)are available.}
}
\value{
RDS or CSV files for each data frame in the list.
}
\description{
Export a list of data frames
}
