% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_calorimetry.R
\name{load_calorimetry}
\alias{load_calorimetry}
\title{Load calorimetry files}
\usage{
load_calorimetry(path = NULL, col_names = c("study_name", "id",
  "session"), file_ext = ".txt", ...)
}
\arguments{
\item{path}{Path to files, Default: NULL}

\item{col_names}{Names for the columns. Default: c("study_name", "id", "session")}

\item{file_ext}{The file extension of your files. Default: ".dat"}

\item{...}{extra parameters to pass to tidyr::separate}
}
\value{
A dataframe.
}
\description{
This function will load all calorimeter files in the same directory,
 into R as a dataframe.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[stringr]{str_match_all}},\code{\link[stringr]{str_subset}}
 \code{\link[purrr]{map_df}},\code{\link[purrr]{set_names}}
 \code{\link[tidyr]{unnest}}
}
