% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_excel.R
\name{load_excel}
\alias{load_excel}
\title{Load data from Excel}
\usage{
load_excel(path = NULL, skip_sheets = NULL)
}
\arguments{
\item{path}{Path to your Excel file, Default: NULL}

\item{skip_sheets}{A vector with the names of sheets you do not want to load, Default: NULL}
}
\value{
One dataframe for each sheet in the original Excel file
}
\description{
This function will load all sheets from an Excel file using readxl.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{excel_sheets}},\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{setdiff}},\code{\link[dplyr]{mutate}}
 \code{\link[purrr]{map}}
 \code{\link[tidyr]{unnest}}
}
