% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rat_eeg.R
\name{load_rat_eeg}
\alias{load_rat_eeg}
\title{Load sleep rat sleep scoring (Spike2)}
\usage{
load_rat_eeg(path = NULL, col_names = c("study", "id", "condition"),
  drop_cols = TRUE, file_ext = ".txt", ...)
}
\arguments{
\item{path}{Path to files as a string, Default: NULL}

\item{col_names}{Names for the columns corresponding to the information on the
filename, Default: c("study", "id", "condition")}

\item{drop_cols}{The Spike2 script will produce files with columns (e.g. T1EEG_PR)
for each electrode and animal recorded. Since usually the information is the
same across all electrodes Default: TRUE}

\item{file_ext}{The extension of the files you want to import, Default: '.txt'}

\item{...}{extra parameters to pass to tidyr::separate}
}
\value{
A data-frame.
}
\description{
Load all .txt files with sleep scoring done with Spike2.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[fs]{dir_ls}},\code{\link[fs]{path_ext_remove}},\code{\link[fs]{path_file}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{select_if}},\code{\link[dplyr]{matches}}
 \code{\link[tidyr]{separate}},\code{\link[tidyr]{unnest}},\code{\link[tidyr]{gather}}
 \code{\link[purrr]{map}}
 \code{\link[readr]{read_table2}}
}
