% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medpsytueR-package.R
\docType{package}
\name{medpsytueR}
\alias{medpsytueR}
\alias{medpsytueR-package}
\title{\code{medpsytueR} package}
\description{
Commonly used functions in the Institute of Medical Psychology, Tuebingen
}
\details{
See the README on
\href{https://github.com/medpsytuebingen/medpsytueR#readme}{GitHub}
}
