% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pvt.R
\name{load_pvt}
\alias{load_pvt}
\title{Load PVT reaction time task}
\usage{
load_pvt(path = NULL, col_names = c("study_name", "id", "session"),
  file_ext = ".dat", ...)
}
\arguments{
\item{path}{Path to files, Default: NULL}

\item{col_names}{Names for the columns. Default: c("study_name", "id", "session")}

\item{file_ext}{The file extension of your files. Default: ".dat"}

\item{...}{extra parameters to pass to tidyr::separate}
}
\value{
A dataframe
}
\description{
Load all .dat files from the Psychomotor Vigilance Task (ePrime) in the same directory.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map_df}},\code{\link[purrr]{set_names}}
 \code{\link[stringr]{str_subset}}
 \code{\link[tidyr]{unnest}}
 \code{\link[fs]{dir_ls}}\code{\link[fs]{path_ext_remove}}
}
